requirements.txt
requests==2.31.0
beautifulsoup4==4.12.3
playwright==1.44.0
Pillow==10.3.0
imagehash==4.3.1
numpy==1.26.4
scikit-learn==1.4.2
sentence-transformers==2.7.0
anthropic==0.28.0
reportlab==4.2.0
pydantic==2.7.1
pydantic-settings==2.2.1
structlog==24.2.0
tenacity==8.3.0
aiohttp==3.9.5
aiofiles==23.2.1
colorlog==6.8.2
python-dotenv==1.0.1
click==8.1.7
tqdm==4.66.4
urllib3==2.2.1
lxml==5.2.2
cssselect==1.2.0
instaloader==4.12
